Adding the calculator read velocity solution plus testing project

+++ CRV.CoreComponent/VelocityRequired.cs
@@ -6,12 +6,12 @@ namespace CRV.CoreComponent
 {
     public class VelocityRequired
     {
-        public int SessionRequired { get; }
+        public double SessionRequired { get; }
         public double PomodorsRequired { get; }
 
         public TimeRequired TimeRequired { get; }
 
-        public VelocityRequired(TimeRequired timeRequired,int sessionRequired, double pomodoroRequired)
+        public VelocityRequired(TimeRequired timeRequired,double sessionRequired, double pomodoroRequired)
         {
             SessionRequired = sessionRequired;
             TimeRequired = timeRequired;
